services/base-service-template/requirements.txt
# Cinema Service - Base Template Requirements

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis[hiredis]>=5.0.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.2

# Cinema Common Library (local development)
-e ../../shared/python-common

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.8
mypy>=1.7.1
